+++ backend/src/users/users.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Put, Delete, Body, Param, UseGuards } from '@nestjs/common';
+import { Controller, Get, Post, Put, Delete, Body, Param, UseGuards, Request } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { JwtAuthGuard } from '../common/guards/jwt-auth.guard';
 import { RolesGuard } from '../common/guards/roles.guard';
@@ -22,17 +22,17 @@ export class UsersController {
   }
 
   @Post()
-  create(@Body() body: CreateUserDto) {
-    return this.usersService.create(body);
+  create(@Body() body: CreateUserDto, @Request() req) {
+    return this.usersService.create(body, { userId: req.user.sub, username: req.user.username });
   }
 
   @Put(':id')
-  update(@Param('id') id: string, @Body() body: UpdateUserDto) {
-    return this.usersService.update(+id, body);
+  update(@Param('id') id: string, @Body() body: UpdateUserDto, @Request() req) {
+    return this.usersService.update(+id, body, { userId: req.user.sub, username: req.user.username });
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.usersService.remove(+id);
+  remove(@Param('id') id: string, @Request() req) {
+    return this.usersService.remove(+id, { userId: req.user.sub, username: req.user.username });
   }
 }
